테이블 스키마 변경 (Integer -> Float)

+++ src/main/java/com/dogfeetbirdfeet/ivorypilatesbackend/dto/schema/CusReg.java
@@ -16,8 +16,8 @@ public class CusReg {
 	private String acctId;
 	private Long grpCusId;
 	private Long clsPassId;
-	private Integer height;
-	private Integer weight;
+	private Float height;
+	private Float weight;
 	private String disease;
 	private String surHist;
 	private String bodyCheckImg;
